+++ src/builtins/commandline.rs
@@ -22,6 +22,7 @@ use crate::wcstringutil::join_strings;
 use std::ops::Range;
 
 /// Which part of the comandbuffer are we operating on.
+#[derive(Eq, PartialEq)]
 enum TextScope {
     String,
     Job,
@@ -103,6 +104,7 @@ fn replace_part(
 fn write_part(
     parser: &Parser,
     range: Range<usize>,
+    range_is_single_token: bool,
     cut_at_cursor: bool,
     token_mode: Option<TokenMode>,
     buffer: &wstr,
@@ -121,19 +123,8 @@ fn write_part(
         return;
     };
 
-    let buff = &buffer[range];
-    let mut tok = Tokenizer::new(buff, TOK_ACCEPT_UNFINISHED);
     let mut args = vec![];
-    while let Some(token) = tok.next() {
-        if cut_at_cursor && token.end() >= pos {
-            break;
-        }
-        if token.type_ != TokenType::string {
-            continue;
-        }
-
-        let token_text = tok.text_of(&token);
-
+    let mut add_token = |token_text: &wstr| {
         match token_mode {
             TokenMode::Expanded => {
                 const COMMANDLINE_TOKENS_MAX_EXPANSION: usize = 512;
@@ -175,7 +166,26 @@ fn write_part(
                 args.push(Completion::from_completion(unescaped));
             }
         }
-    }
+    };
+
+    let buff = &buffer[range];
+    if range_is_single_token {
+        add_token(buff);
+    } else {
+        let mut tok = Tokenizer::new(buff, TOK_ACCEPT_UNFINISHED);
+        while let Some(token) = tok.next() {
+            if cut_at_cursor && token.end() >= pos {
+                break;
+            }
+            if token.type_ != TokenType::string {
+                continue;
+            }
+
+            let token_text = tok.text_of(&token);
+            add_token(token_text);
+        }
+    };
+
     for arg in args {
         streams.out.appendln(arg.completion);
     }
@@ -642,6 +652,7 @@ pub fn commandline(parser: &Parser, streams: &mut IoStreams, args: &mut [&wstr])
         write_part(
             parser,
             range,
+            buffer_part == TextScope::Token,
             cut_at_cursor,
             token_mode,
             current_buffer,