+++ src/highlight/highlight.rs
@@ -1,8 +1,9 @@
 //! Functions for syntax highlighting.
 use crate::abbrs::{self, with_abbrs};
 use crate::ast::{
-    self, Argument, Ast, BlockStatement, BlockStatementHeaderVariant, DecoratedStatement, Keyword,
-    Leaf, List, Node, NodeVisitor, Redirection, Token, Type, VariableAssignment,
+    self, Argument, Ast, BlockStatement, BlockStatementHeaderVariant, BraceStatement,
+    DecoratedStatement, Keyword, Leaf, List, Node, NodeVisitor, Redirection, Token, Type,
+    VariableAssignment,
 };
 use crate::builtins::shared::builtin_exists;
 use crate::color::RgbColor;
@@ -869,6 +870,9 @@ impl<'s> Highlighter<'s> {
             ParseTokenType::end | ParseTokenType::pipe | ParseTokenType::background => {
                 role = HighlightRole::statement_terminator
             }
+            ParseTokenType::left_brace | ParseTokenType::right_brace => {
+                role = HighlightRole::keyword;
+            }
             ParseTokenType::andand | ParseTokenType::oror => role = HighlightRole::operat,
             ParseTokenType::string => {
                 // Assume all strings are params. This handles e.g. the variables a for header or
@@ -1063,6 +1067,12 @@ impl<'s> Highlighter<'s> {
         self.visit(&block.end);
         self.pending_variables.truncate(pending_variables_count);
     }
+    fn visit_brace_statement(&mut self, brace_statement: &BraceStatement) {
+        self.visit(&brace_statement.left_brace);
+        self.visit(&brace_statement.args_or_redirs);
+        self.visit(&brace_statement.jobs);
+        self.visit(&brace_statement.right_brace);
+    }
 }
 
 /// Return whether a string contains a command substitution.
@@ -1121,6 +1131,7 @@ impl<'s, 'a> NodeVisitor<'a> for Highlighter<'s> {
                 self.visit_decorated_statement(node.as_decorated_statement().unwrap())
             }
             Type::block_statement => self.visit_block_statement(node.as_block_statement().unwrap()),
+            Type::brace_statement => self.visit_brace_statement(node.as_brace_statement().unwrap()),
             // Default implementation is to just visit children.
             _ => self.visit_children(node),
         }